python-server/requirements.txt
torch==2.0.0
torchvision==0.15.1
flask==2.3.0
python-dotenv==0.21.0
boto3==1.28.30
uvicorn==0.23.0